small tidy ups
