+++ proxy/nginx.conf.template
@@ -25,10 +25,6 @@ http {
 
     keepalive_timeout  65;
 
-    # this sets the maximum size information to be unlimited;
-    # TODO: We might want to revise this!
-    client_max_body_size 0;
-
     # Main server configuration. See below for redirects.
     server {
         listen       443 ssl;
