+++ src/daedalus_config.py
@@ -1,0 +1,72 @@
+import time
+import os
+
+import docker
+import json
+import constellation
+import constellation.config as config
+
+
+class DaedalusConfig:
+    def __init__(self, path, config_name=None, options=None):
+        dat = config.read_yaml("{}/daedalus.yml".format(path))
+        dat = config.config_build(path, dat, config_name, options=options)
+        self.path = path
+        self.data = dat
+        self.vault = config.config_vault(dat, ["vault"])
+        self.network = config.config_string(dat, ["docker", "network"])
+        self.container_prefix = config.config_string(dat, ["docker", "prefix"])
+
+        self.containers = {
+            "api": "api",
+            "web-app-db": "web-app-db",
+            "web-app": "web-app",
+            "proxy": "proxy"
+        }
+
+        self.volumes = {
+            "daedalus-data": "daedalus-data",
+            "proxy-logs": "proxy-logs"
+        }
+
+        # api
+        self.api_ref = self.get_image_reference("api", dat)
+        self.api_port = config.config_integer(dat, ["api", "port"])
+
+        # web_app_db
+        self.web_app_db_ref = self.get_image_reference("web_app_db", dat)
+        self.web_app_db_port = config.config_integer(dat, ["web_app_db", "port"])
+        self.web_app_db_data_location = config.config_string(dat, ["web_app_db", "data_location"])
+
+        # web_app
+        self.web_app_ref = self.get_image_reference("web_app", dat)
+        self.web_app_port = config.config_integer(dat, ["web_app", "port"])
+
+        # proxy
+        self.proxy_ref = self.get_image_reference("proxy", dat)
+        self.proxy_host = config.config_string(dat, ["proxy", "host"])
+        self.proxy_port_http = config.config_integer(dat, ["proxy", "port_http"])
+        self.proxy_port_https = config.config_integer(dat, ["proxy", "port_https"])
+        self.proxy_logs_location = config.config_string(dat, ["proxy", "logs_location"])
+
+        if "ssl" in dat["proxy"]:
+            self.proxy_ssl_certificate = config.config_string(dat,
+                                                              ["proxy",
+                                                               "ssl",
+                                                               "certificate"])
+            self.proxy_ssl_key = config.config_string(dat,
+                                                      ["proxy",
+                                                       "ssl",
+                                                       "key"])
+            self.ssl = True
+        else:
+            self.ssl = False
+
+    def get_image_reference(self, config_section, dat):
+        repo = config.config_string(
+                    dat, [config_section, "image", "repo"])
+        name = config.config_string(
+            dat, [config_section, "image", "name"])
+        tag = config.config_string(
+            dat, [config_section, "image", "tag"])
+        return constellation.ImageReference(repo, name, tag)
